ProyectoDjango/requirements.txt
Django>=4.0
gunicorn
mysqlclient
django-environ
whitenoise
django-ckeditor
Pillow>=9.0.0
boto3
django-storages